# graphqlbackend: Use pointers to zoekt.Repository instead of copying