With large numbers of repositories, this map can grow really big and put
a lot of pressure on the GC. Instead of copying the structs, use
pointers.

Test plan: go test ./...