@mrnugget 2019-06-25
@keegancsmith Just pushed another commit that adds a test for the `zoektIndexedRepos` function. @tsenart and I paired on this and we decided to add a test after doing some more optimizations and getting less and less confident that we didn't break anything.