@tsenart 2019-06-25
I'd like to see the [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat) output of this change.