@mrnugget 2019-06-25
Here is the `benchstat` output:

```
λ benchstat old.txt new.txt
name                  old time/op    new time/op    delta
_zoektIndexedRepos-8    4.89ms ± 2%    3.37ms ± 6%  -31.09%  (p=0.000 n=8+10)

name                  old alloc/op   new alloc/op   delta
_zoektIndexedRepos-8    4.10MB ± 0%    2.50MB ± 0%  -39.01%  (p=0.000 n=9+9)

name                  old allocs/op  new allocs/op  delta
_zoektIndexedRepos-8     30.1k ± 0%     20.1k ± 0%  -33.27%  (p=0.000 n=10+10)
```